+++ multiboot2/CHANGELOG.md
@@ -8,6 +8,7 @@
 - `BootInformation::elf_sections` is now deprecated and replaced by the newly
 - added `BootInformation::elf_sections_tag`. On the returned type, you can call
   `.sections()` to iterate the sections
+- Fixed the debug output of `BootInformation`
 
 ## v0.23.0 (2024-09-17)